+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsingestor/RtdMsIngestorApplication.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.rtd.ms.rtdmsingestor;
 
-import it.gov.pagopa.rtd.ms.rtdmsingestor.configuration.WalletConfiguration;
+import it.gov.pagopa.rtd.ms.rtdmsingestor.configuration.properties.WalletConfigurationProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -9,7 +9,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
  * RTD microsrvice to ingest on a Kafka Topic decrypted files contsinint electronic payments.
  */
 @SpringBootApplication
-@EnableConfigurationProperties(WalletConfiguration.class)
+@EnableConfigurationProperties(WalletConfigurationProperties.class)
 public class RtdMsIngestorApplication {
   /**
    * Main method to run the Ingestor Miscroservice.
