+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsingestor/adapter/ContractAdapter.java
@@ -3,7 +3,6 @@ package it.gov.pagopa.rtd.ms.rtdmsingestor.adapter;
 import it.gov.pagopa.rtd.ms.rtdmsingestor.model.WalletContract;
 import org.springframework.stereotype.Service;
 
-@Service
 public class ContractAdapter {
 
   public WalletContract adapt(WalletContract contract) {
