+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsingestor/utils/Anonymizer.java
@@ -1,28 +1,33 @@
 package it.gov.pagopa.rtd.ms.rtdmsingestor.utils;
 
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.util.Base64;
-import javax.crypto.Mac;
-import javax.crypto.spec.SecretKeySpec;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.codec.binary.Hex;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import javax.crypto.Mac;
+import javax.crypto.spec.SecretKeySpec;
+import java.security.InvalidKeyException;
+import java.security.NoSuchAlgorithmException;
+import java.util.Base64;
+
 @Service
 @Slf4j
-@RequiredArgsConstructor
 public class Anonymizer {
 
   private static final String FAILED_HMAC_STUB = "###FAILED_TO_PRODUCE_HMAC###";
   private static final String ALGORITHM = "HmacSHA256";
 
-  @Value("${ingestor.contractIdObfuscationHmacKey}")
-  private String secretKey;
 
-  /**
+  private final String secretKey;
+
+    public Anonymizer(
+            @Value("${ingestor.contractIdObfuscationHmacKey}") String secretKey
+    ) {
+        this.secretKey = secretKey;
+    }
+
+    /**
    * This method anonymize a given string with a determined algorithm and a key recovered from
    *
    * @param toAnonymize string to anonymize
